code smell billing
