+++ apps/frontend/src/components/billing/types.ts
@@ -1,0 +1,16 @@
+interface Tiers {
+    month: Array<{
+        name: 'Pro' | 'Standard';
+        recurring: 'month' | 'year';
+        price: number;
+    }>;
+    year: Array<{
+        name: 'Pro' | 'Standard';
+        recurring: 'month' | 'year';
+        price: number;
+    }>;
+}
+
+export {
+    Tiers
+}
